+++ dune/localfunctions/lagrange/pk2d.hh
@@ -11,6 +11,8 @@
 #include <dune/localfunctions/common/localtoglobaladaptors.hh>
 #include <dune/localfunctions/lagrange/lagrangesimplex.hh>
 
+#warning This header is deprecated
+
 namespace Dune
 {
 
@@ -19,7 +21,9 @@ namespace Dune
       \deprecated This class is obsolete. Please use LagrangeSimplexLocalFiniteElement instead!
    */
   template<class D, class R, unsigned int k>
-  using Pk2DLocalFiniteElement = LagrangeSimplexLocalFiniteElement<D,R,2,k>;
+  using Pk2DLocalFiniteElement
+    [[deprecated("use LagrangeSimplexLocalFiniteElement instead")]]
+    = LagrangeSimplexLocalFiniteElement<D,R,2,k>;
 
 
   //! Langrange finite element of arbitrary order on triangles